+++ src/Depressurizer.Core/Interfaces/IGameList.cs
@@ -18,7 +18,7 @@ namespace Depressurizer.Core.Interfaces
 
         List<Filter> Filters { get; set; }
 
-        Dictionary<int, GameInfo> Games { get; set; }
+        Dictionary<long, GameInfo> Games { get; set; }
 
         #endregion
 
@@ -28,15 +28,15 @@ namespace Depressurizer.Core.Interfaces
 
         Filter AddFilter(string name);
 
-        void AddGameCategory(int appId, Category category);
+        void AddGameCategory(long appId, Category category);
 
-        void AddGameCategory(int[] appIds, Category category);
+        void AddGameCategory(long[] appIds, Category category);
 
         bool CategoryExists(string prefix);
 
-        void ClearGameCategories(int appId, bool preserveFavorite);
+        void ClearGameCategories(long appId, bool preserveFavorite);
 
-        void ClearGameCategories(int[] appIds, bool preserveFavorite);
+        void ClearGameCategories(long[] appIds, bool preserveFavorite);
 
         void ExportSteamConfig(long steamId, bool discardMissing, bool includeShortcuts);
 
@@ -48,33 +48,33 @@ namespace Depressurizer.Core.Interfaces
 
         Filter GetFilter(string acgFilter);
 
-        void HideGames(int appId, bool isHidden);
+        void HideGames(long appId, bool isHidden);
 
-        void HideGames(int[] appIds, bool isHidden);
+        void HideGames(long[] appIds, bool isHidden);
 
-        int ImportSteamConfig(long steamId, SortedSet<int> ignore, bool includeShortcuts);
+        int ImportSteamConfig(long steamId, SortedSet<long> ignore, bool includeShortcuts);
 
-        int IntegrateGameList(GetOwnedGamesObject doc, bool overwrite, SortedSet<int> ignore, out int newItems);
+        int IntegrateGameList(GetOwnedGamesObject doc, bool overwrite, SortedSet<long> ignore, out int newItems);
 
-        int IntegrateGameList(IXPathNavigable doc, bool overwrite, SortedSet<int> ignore, out int newItems);
+        int IntegrateGameList(IXPathNavigable doc, bool overwrite, SortedSet<long> ignore, out int newItems);
 
         bool RemoveCategory(Category category);
 
         int RemoveEmptyCategories();
 
-        void RemoveGameCategory(int appId, Category category);
+        void RemoveGameCategory(long appId, Category category);
 
-        void RemoveGameCategory(int[] appIds, Category category);
+        void RemoveGameCategory(long[] appIds, Category category);
 
         Category RenameCategory(Category category, string newName);
 
-        void SetGameCategories(int[] appIds, Category cat, bool preserveFavorites);
+        void SetGameCategories(long[] appIds, Category cat, bool preserveFavorites);
 
-        void SetGameCategories(int appId, ICollection<Category> catSet, bool preserveFavorites);
+        void SetGameCategories(long appId, ICollection<Category> catSet, bool preserveFavorites);
 
-        void SetGameCategories(int[] appIds, ICollection<Category> catSet, bool preserveFavorites);
+        void SetGameCategories(long[] appIds, ICollection<Category> catSet, bool preserveFavorites);
 
-        int UpdateGameListFromOwnedPackageInfo(long accountId, SortedSet<int> ignored, out int newApps);
+        int UpdateGameListFromOwnedPackageInfo(long accountId, SortedSet<long> ignored, out int newApps);
 
         #endregion
     }
