+++ src/Depressurizer/AutoCats/CDlgCurator.cs
@@ -15,7 +15,7 @@ namespace Depressurizer.AutoCats
     {
         #region Fields
 
-        public Dictionary<int, CuratorRecommendation> CuratorRecommendations;
+        public Dictionary<long, CuratorRecommendation> CuratorRecommendations;
 
         public int TotalCount;
 
@@ -29,7 +29,7 @@ namespace Depressurizer.AutoCats
         {
             SetText(GlobalStrings.CDlgCurator_GettingRecommendations);
             this.curatorId = curatorId;
-            CuratorRecommendations = new Dictionary<int, CuratorRecommendation>();
+            CuratorRecommendations = new Dictionary<long, CuratorRecommendation>();
         }
 
         #endregion
@@ -99,9 +99,9 @@ namespace Depressurizer.AutoCats
         ///     https://store.steampowered.com/curator/{0}/ajaxgetfilteredrecommendations/render/?query=&amp;start={1}&amp;count=50
         /// </param>
         /// <returns>A dictionary containing ids of games and their respective recommendations</returns>
-        private static Dictionary<int, CuratorRecommendation> GetCuratorRecommendationsFromPage(string page)
+        private static Dictionary<long, CuratorRecommendation> GetCuratorRecommendationsFromPage(string page)
         {
-            Dictionary<int, CuratorRecommendation> curatorRecommendations = new Dictionary<int, CuratorRecommendation>();
+            Dictionary<long, CuratorRecommendation> curatorRecommendations = new Dictionary<long, CuratorRecommendation>();
             Regex curatorRegex = new Regex(@"data-ds-appid=\""(\d+)\"".*?color_(\w+)", RegexOptions.Singleline | RegexOptions.Multiline | RegexOptions.Compiled);
             MatchCollection matches = curatorRegex.Matches(page);
             if (matches.Count > 0)
